requirements.txt
apache-airflow-providers-google==10.3.0